Print out "your review request still exists" when there are field errors.
When posting a change, if the diff has a parse error, we printed out a message
that a review request was created, including the URL to it. In the case where
updating the draft fails (for example, when assigning reviewers based on command-line
arguments), we weren't showing a similar error. This change fixes it so that
the user gets the full error.

Testing done:
Tried to post a change with an invalid group listed in --target-groups. Saw an
error that included the new output.

Fixes bug 3428.

Reviewed at https://reviews.reviewboard.org/r/6098/