Support authenticating in RBTools with API tokens.
Review Board 2.1+ supports authenticating using API tokens. This change
adds a new --api-token= parameter for commands to make use of these when
authenticating.

Since authentication happens before accessing resources, we can't really
check for a capability before deciding whether or not to allow a user to
pass a token. Passing a token to a server that doesn't support it is
harmless, though.

Reviewed at https://reviews.reviewboard.org/r/6237/