Add -C option to commit patches immediately
It is very useful to be able to script a series of patch applications all at
once (such as when reviewing a patch series). In these cases, it is useful to
be able to accept the default commit message for each patch.

Testing Done:
Patched my local RBTools installation with these changes and pulled a patch
into my working tree using the -C option. It applied with the default commit as
expected.

Reviewed at https://reviews.reviewboard.org/r/6122/