Fix handling of deleted files in submitted perforce changelists.
When we fetch the list of affected files in a given changelist, perforce
reports deleted files with a revision number. That revision corresponds to the
deleted version of the file, and p4 print will just return empty. This is
similar to edits, where the revision is actually the post-edited revision
rather than the pre-edited version, and we need to decrement the revision.

Testing done:
- Ran through some p4 deletes.
- Ran unit tests.

Fixes bug 3497.

Reviewed at https://reviews.reviewboard.org/r/6191/