Fix posting with SVN.
The recent patch I accepted to reintroduce usage of find_server_repository_info
exposed a bug. The SVN implementation of this method hadn't been updated to
cope with recent changes to the attribute behavior of API objects, and so
trying to use repository.get('mirror_path') caused an AttributeError. We can
blame Python's dumb slots for that.

I've rewritten this part of the code to avoid the built-in get() method and
check hasattr instead.

Testing done:
Tried to post a change against my local devserver with an svn repository.
Before this change, I'd always get an AttributeError when trying to find the
repository. After, everything posts fine.

Reviewed at https://reviews.reviewboard.org/r/6285/