Add -X support for the svn SCMClient.
The svn backend now supports excluding files through a
filterdiff-esque method: `SVNClient._filter_diff`. This
method looks through the output of `svn diff` for lines
that begin with `Index:` that mark the beginning of a
new file in the diff. If the file matches any given
patterns, it will not yield the lines corresponding to
that file's diff.

Testing Done:

Ran `rbt diff -X ...` in an SVN repository and successfully
excluded files and empty files.

Reviewed at https://reviews.reviewboard.org/r/6345/