Git client: properly handle deleted files when making SVN-compatible diff
When deleting a file, git-diff outputs "+++ /dev/null" and "deleted file mode",
neither of which are understood by SVN. Rather, SVN expects the new filename to
be the same as the deleted one. This is important when doing file renames or
deletions when using git-svn.

Testing Done:
Successfully used this code to post a review for a basic set of file renames
(~10 files). Posted to RB 1.6 and it seemed to be happy.

Reviewed at https://reviews.reviewboard.org/r/6396/