Add an installer for MacOS X.
This adds an installer for RBTools on MacOS X, which bundles the command
line utilities and the modules for both Python 2.6 and 2.7, both of
which are standard on modern versions of MacOS X.

The installer is built with the build-mpkg.sh script, which will
generate the appropriate builds for Python 2.6 and 2.7 versions, and
assemble all the pieces into a standard .pkg file that can be
distributed.

Currently, attempting to install this package will result in a complaint
about it not being signed, making it difficult and scary to install.
This will be fixed when we have the ability to sign the package, which
will happen prior to the package's official release.

Reviewed at https://reviews.reviewboard.org/r/6401/