Fix GIT_USE_EXT_DIFF.
The code to enable-disable external diff programs for git wasn't quite right,
in that it worked as intended if the config parameter was not present, but if
it was, the value was backwards.

Testing done:
Set my diff viewer to an external program and verified that the setting now
worked as intended.

Reviewed at https://reviews.reviewboard.org/r/6417/