Don't fail when using --diff-filename=- and credentials on the command line.
When passing a diff in on stdin, we can't also prompt for credentials.
Unfortunately, the check for this was also making it so that it would fail
even in the case where credentails were passed in on the command-line. Moving
the check fixes the problem.

Testing done:
- Smoke tested --diff-filename.
- Ran unit tests.

Fixes bug 3313.

Reviewed at https://reviews.reviewboard.org/r/6538/