This introduces the Windows installer that we'll be shipping with official RBTools releases, starting in 0.7. It's based upon Bruce Cran's installer work, which makes use of WiX and Portable Python to produce a standalone RBTools install independent of any system installs of Python.
This change extends Bruce's work to automate the building of the installer
and to isolate the build environment from the system. There's a
build-installer.bat file that handles the installation of Portable Python
(guiding the user through a Minimal installation on the first build),
stripping it, installing RBTools and its dependencies, and building the
WiX installer.

This must run on a Windows install, and does not run automatically when
we perform builds. It must instead be run manually within a suitable
Windows build environment. That environment does not need Python installed
up-front, but does need WiX and MSBuild.exe (part of Visual Studio or
MS Build Tools, though currently only tested with the former).

Reviewed at https://reviews.reviewboard.org/r/6690/