Added 'rbt land', a command for landing reviewed changes.
rbt land is similar in concept to rbt patch, in that it applies a diff from
a review request and creates a commit for it. It goes a few steps further,
though.

Before applying the diff, it checks that the change has been properly
reviewed. On Review Board 2.0+, this checks the review request's approval
state, which extensions can influence. On older versions, it just checks for
one or more Ship It!'s.

Once the change has been approved and applied, it is committed to the
target landing branch, and then pushed upstream.

Unlike rbt patch, either a review request ID or a local branch can be landed.
This makes the command useful both for developers pushing their own changes,
and developers pushing contributions from other contributors.

Reviewed at https://reviews.reviewboard.org/r/6509/