Fix losing the commit summary when using rbt stamp.
When rebuilding the commit message, rbt stamp was losing the summary.
This was due to the code only factoring in the description for the new
message.

We now build upon the raw commit message, rather than a parsed message.
We also don't try to stamp the commit if it's already been stamped.

Reviewed at https://reviews.reviewboard.org/r/6701/