Support guessing review requests in rbt land.
rbt land now defaults to guessing review requests, unless using -r.
It will attempt to find a review request matching the specified branch,
or the current branch if one was not specified. It will prompt for a
match if unsure.

If a match is found, it will be used for the resulting commit message in
the merge.

Reviewed at https://reviews.reviewboard.org/r/6713/