Default to --local in rbt land if merging a branch.
Currently, merging a local branch requires specifying --local. This
ended up being a bit more work than wanted for the average user.

Now, --local is the default unless both -r is used and a specific branch
is not specified. This means that if you're just landing a remote change
using -r, it will land correctly as before, but if you're trying to land
the current branch (without specifying -r) or landing a specified
branch, it will assume it's local and doesn't need to be patched.

Reviewed at https://reviews.reviewboard.org/r/6714/