Fix issues landing remote patches using rbt land.
rbt land made some assumptions that resulted in problems with landing
remote branches. Certain operations, such as comparing the current and
destination branches, and merging branches, assumed we were landing
local branches in all cases, but this wasn't so with remote patches.
We now only execute some of that logic if we're landing local branches.

Patching was also broken, as it was only patching to the index and not
to a commit. We now pass -C to fix this.

Reviewed at https://reviews.reviewboard.org/r/6715/