Fix posting with utf-8 changes.
This fixes a regression from the UTF-8 changes that I made recently. In a few
places, we were passing unicode objects in for the HTTP method or headers. For
the most part, this was fine because internally, HTTPRequest would concatenate
everything into one big unicode string and then encode it as ascii before
sending it on the wire. Unfortunately, if the posted changes (or description)
included any unicode characters, it would try to encode as ascii and fail.
Keeping everything that goes to httplib as bytes will avoid this.

Testing done:
Successfully posted a change that included utf-8 characters in the diff.