Fix issues guessing review requests in rbt land.
rbt land wasn't including the specified branch name as part of the list
of arguments that get passed to get_revisions(). This caused the review
request guessing code to look at HEAD, instead of the specified branch.

That bug exposed another problem, which is that the review request
guessing code didn't pre-define the summary and description, leading to
an error later on if a summary or description was not found. We now
properly bail in this case.

Reviewed at https://reviews.reviewboard.org/r/6726/