Add a --dry-run argument to rbt land, and operations output.
rbt land is a command that can have some real impact on a source tree,
so it's important to let people feel comfortable knowing what it's going
to do.

This change adds a `--dry-run` argument that simulates the entire process
of landing a change, merging the branch, and pushing. It won't actually
perform those operations, but will act as if it did. This means that
some validation steps performed by the SCM tool may not occur, but will
still give a good sense of what's going to happen.

It also adds some output indicating what's going to happen at each
important stage, so that if something goes wrong, the user has a good
idea as to where and with what branches.

Testing Done:
Tested with remote changes and local branches, with both squashing and merging.

Verified that the tree was untouched, but that all the expected operations were
being "performed" (as per the output).

Reviewed at https://reviews.reviewboard.org/r/6719/