Add an option for invoking an editor for the commit in rbt land.
rbt land now has a -e (--edit) option for invoking an editor to edit the
commit message before landing a change. The message will be the
generated commit containing the "Reviewed at" line.

This gives the committer the ability to reword the message from a
contributor, or to add additional information like bugs closed.

Testing Done:
Landed a local branch using -e, and was presented with my editor for the
commit message. Changed the commit, saved and exited, and saw that the commit
was created with my new message.

Reviewed at https://reviews.reviewboard.org/r/6722/