Improve usage of some rbt land options.
This adds short flags for the --push (-p) and --no-push (-n) options,
which are likely to be some of the more commonly-used options.

It adds a configuration key to set the default for LAND_SQUASH, and adds
a --no-squash option to override if set.

It also adds a missing period in the help text for --squash.

Testing Done:
Tested --squash and --no-squash.

Tested with -p and -n.

Reviewed at https://reviews.reviewboard.org/r/6723/