Improve the wording and metadata for RBTools command options.
Many of the older options for the RBTools commands had help text that
was less than clear, and inferior to the wording in our documentation.
In some cases, we just had too much text in one big paragraph, and in
others, the options weren't as clear as they could have been.

The wording for some options have been improved, and newlines have been
added to break apart paragraphs. This works along with a custom help
text formatter, which preserves paragraphs (as the default ones strip
them away).

Some additional metadata was also added to several options for extended
help descriptions (matching extra information currently in the docs) and
to provide meta variables and versioning information. This sets us up
for the next change, which auto-generates much of the RBTools doc pages
based on the information in the code, helping to keep the docs in-sync
with the current state of the world.

Reviewed at https://reviews.reviewboard.org/r/6710/