Add an option to delete branches after landing them.
This introduces --delete-branch and --no-delete-branch options for
handling whether a branch should be deleted after successfully being
landed.

By default, a branch is deleted, but LAND_DELETE_BRANCH can be set
in .reviewboardrc to change this behavior.

Testing Done:
Tested landing a branch without --delete-branch. Saw that the branch was
deleted after landing.

Tested landing with --no-delete-branch. Saw that the branch was not deleted
after landing.

Changed the setting in .reviewboardrc and tested. Saw that the branch was not
deleted.

Reviewed at https://reviews.reviewboard.org/r/6727/