Fix a regression in landing the current branch.
A fix I made for landing a specified branch broke landing the current
branch. This broke because the new code was unconditionally placing the
branch name parameter into the command arguments list, but if
unspecified, it was placing None in that list, which wasn't a valid
value. We now conditionally insert it.

Testing Done:
Landed the current branch successfully.

Reviewed at https://reviews.reviewboard.org/r/6728/