Sign the Windows installer with our certificate.
The Windows installer is now signed with our certificate, guaranteeing
that users will see "Beanbag, Inc." instead of "Unknown publisher" when
installing the package. This helps people to know that they're getting
an official installer, and that it's safe to install.

To build a signed installer, our official certificate and private key
must be installed on the machine. We do not bundle those, for obvious
reasons.

Testing Done:
Built an installer and ran it. Saw that it said Beanbag, Inc.

Copied the installer to a VM that didn't have the certificate and private
key installed. Ran the installer and saw the same correct certificate info.

Reviewed at https://reviews.reviewboard.org/r/6730/