Use a timestamp server for the RBTools Windows installer.
This uses the Comodo timestamp server when signing the Windows installer
using our authenticode certificate. This ensures that when the
certificate expires, the installers will still be functional, since
there will be a way to ensure that the certificate was valid when the
installer was made.

Testing Done:
Built the installer. I didn't have any errors.

Set the timestamp server to google.com, just to verify that it was being
used for timestamping. Saw an error.

Reviewed at https://reviews.reviewboard.org/r/6749/