Reduce the number of API requests when posting review requests.
When posting review requests, we have two unnecessary API requests: One
to the top-level validation resource (used solely as a link to the diff
validation resource), and one for the user source coming off of a
session resource.

We now use the URI templates to get the diff validation resource
directly, and we expand the user resource in the session resource to
prevent the need to fetch that request.

Testing Done:
Posted this change. No problems.

Saw the debug output and verified we weren't making those two requests.

Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/6844/