Fixed remote branch discovery through subgit
Git-svn checks the output of 'git svn rebase -n' to determine which
SVN branch the current git branch is based on.
In subgit, we crawl up through parent commits until we find one that
subgit knows about (one that is in its notes), and we call that
the upstream branch.
This also introduces a git config option, rbtools.subgit.svn_url,
that can be set if you don't have ssh access to the subgit server.
This should probably be 1) documented; and 2) in .reviewboardrc instead.