Respect --edit for `rbt land` with remote changes.
When using `rbt land` to apply a remote change from a review request, we
weren't respecting the --edit flag, which meant that if a user wanted to
reformat the commit message (for example, wrapping to 80 columns), they'd have
to amend the commit later. I've added a check in the `patch` method to use
either `-C` or `-c` as appropriate.

Testing done:
Used `rbt land -r XXXX -e --dest master` and saw that it invoked my editor.

Reviewed at https://reviews.reviewboard.org/r/7010/