Support for individual commit validation
Refactor the `DiffCommitListResource` into a mixin to allow commmit
behvaiour to be shared between the original `DiffCommitListResource`
and the new `ValidateDiffCommitResource`. This new resource is
responsible for validating a commit (the diff and the metadata) before
it is uploaded and added to the review request.

This new validator is used to validate all commits as they are
uploaded. This replaces the old behaviour where only the diff was
validated (using the `ValidateDiffResource`) but none of the metadata
was.

Testing Done:
Ran unit tests.

Able to upload a review request with history and have its commits
validated.

Reviewed at https://reviews.reviewboard.org/r/7038/