Don't log output from 'diff' commands that return rc != 0.
Many diff commands use a non-zero return code to indicate that the given files
are different. This is fine, except we log the output of commands even when
we're expecting non-zero return codes. This was causing issues when the output
of the diff command wasn't ASCII, and we tried to format it into a logging
call.

We don't actually care about logging the output of diffs, since it makes our
debug output pretty verbose. This change adds a new argument and some plumbing
to make sure that we don't log the output of diff commands.

Testing done:
- Ran unit tests.
- Verified that diff contents were no longer logged with SCMs where they were
  previously (perforce, svn)

Fixes bug 3781.

Reviewed at https://reviews.reviewboard.org/r/7056/