Save response body as binary when updating cache
In the HTTP cache for the API, the `response_body` field is marked as a
BLOB field and can contain arbitrary data. Previously when a cache entry
was updated, the content being inserted into the `response_body` field
was not being marked as binary, which caused issues when dealing with
diffs that had non-ASCII characters. This patch fixes that issue by
ensuring the the response body is being saved as binary when the field
is updated.

A unit test has been added to verify that non-ASCII data can be written
to the HTTP cache without error.

Testing Done:
Ran unit tests.

Verified the supplied unit test fails without this patch applied.

Reviewed at https://reviews.reviewboard.org/r/7098/