Allow rbt patch to work in SVN subdirectories
Previously `rbt patch` would not work in a subdirectory or subdirectory
checkout. This was because the incorrect number of leading directories
were being stripped (due to our SVN diffs having absolute paths). We now
compute the correct number of leading directories to strip when applying
patches and they now apply correctly.

If a patch for SVN is generated in a different directory, files included
in the patch that are not under the current directory will not be
applied. If this occurs, the user will be informed.

This patch also fixes the issue of passing a value `--px` flag on the
commandline causing a formatting error.

Testing Done:
Ran unit tests.

Manually tested the cases that gmyers outlined in
https://reviews.reviewboard.org/r/7095/. `rbt patch` works in all of
these cases.

Reviewed at https://reviews.reviewboard.org/r/7110/