Add revert patch support to the rbt patch command
The `rbt patch` command is now capable of reverting patches via the
`-R`/`--revert` command line flag.

Testing Done:
Ran unit tests.

Tested `rbt patch -R` with CVS, Git, Perforce, and SVN.

Applied a patch with Perforce that had empty files and was able to
revert it with `rbt patch -R`.

Applied a patch with SVN that had empty files and was able to revert
it with `rbt patch -R`.

Running `rbt patch -R` when using a Mercurial repository results in an
error as expected.

I am not able to verify that this works with Plastic or ClearCase, but
it should work because they both use the default `apply_patch` method.

Reviewed at https://reviews.reviewboard.org/r/7084/