Fetch patches from individual commits
The patch fetching behaviour of both the `DiffResource` and the
`FileDiffResource` has been refactored out into the
`BasePatchResource`. This base class is used for the
`DiffCommitResource` to fetch patch files for individual commits.

Testing Done:
Ran unit tests.

Manually printed the contents of a patch from the
`DiffCommitResource` and verified it was correct.

Reviewed at https://reviews.reviewboard.org/r/7133/